requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
Django==3.0.6
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pycparser==2.20
pylint==2.5.2
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
